fix the LRIT MSG (8 bits) calibration
In the case for LRIT MSG, the calibration was erroneous since the counts
are rounded to 8 bits instead of 10. In order for the calibration to work,
the raw counts have to be multiplied by 4 before hand. At the same time,
the data is converted to 16 uints for the multiplication to work.

Signed-off-by: Martin Raspaud <martin.raspaud@smhi.se>